fix(덱): 1021 회전하는 큐
클래스로 만들 필요가 없는 것 같아서 아예 배열 코드로 수정했습니다.
